+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 ## [Unreleased]
+### Removed
+- Shutdown hook register on supplied Launch, by @HardNorth
+
+## [5.0.3]
+### Fixed
+- Backgrounds finish with `FAILED` status, by @HardNorth
+
+## [5.0.2]
 ### Changed
 - Karate dependency marked as `compileOnly` to force users specify their own version of Karate, by @HardNorth
 - Client version updated on [5.2.5](https://github.com/reportportal/client-java/releases/tag/5.2.5), by @HardNorth